requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pydantic==1.10.2
mysql-connector-python==8.0.30  # Для синхронного подключения к MySQL
pytest==7.1.3  # Для тестирования
httpx==0.23.0  # HTTP client,
starlette~=0.20.4
PyJWT~=2.8.0